add 파일 저장 파일 저장 중 에러 시 롤백

+++ src/main/java/com/sample/filestorage/entity/Myfile.java
@@ -10,7 +10,7 @@ public class Myfile {
     private int id;
     private final String originName;
     private final String fName;
-    private final int fSize;
+    private final long fSize;
     private final String fType;
     private Timestamp fDatetime;
     private final String storedPath;
